+++ serialize/serializer_test.ts
@@ -61,6 +61,11 @@ function clone(object: any) {
 const emptyOpt: ModOptions = {
   lang: "fr",
   storyPath: "",
+  imageItemGenFont: "Arial",
+  rssMinDuration: 0,
+  thumbnailFromFirstItem: false,
+  rssSplitLength: 10,
+  skipRssImageDl: false,
 };
 
 Deno.test("serializePack-min", async () => {
@@ -102,6 +107,11 @@ Deno.test("serializePack-next-menu", async () => {
     lang: "fr",
     storyPath: "",
     selectNextStoryAtEnd: true,
+    imageItemGenFont: "Arial",
+    rssMinDuration: 0,
+    thumbnailFromFirstItem: false,
+    rssSplitLength: 10,
+    skipRssImageDl: false,
   });
   assertEquals(
     clearIds(clone(pack)),