+++ test_data/test_data.ts
@@ -54,7 +54,7 @@ export const expectedMinPack: Pack = {
           class: "StageNode-Story",
           audio: "46fe70d98b9763ab70b7e7cea4627e4f8b7c585a.ogg",
           image: null,
-          name: "alice-city.ogg Stage node",
+          name: "alice-city",
           okTransition: null,
           path: undefined,
         },
@@ -117,7 +117,7 @@ export const expectedMinPackSerialized = {
       },
       homeTransition: null,
       image: null,
-      name: "alice-city.ogg Stage node",
+      name: "alice-city",
       okTransition: null,
       position: {
         x: 0,
@@ -237,12 +237,12 @@ export const expectedMoyPack: Pack = {
           imagePath: undefined,
           imageTimestamp: undefined,
           okTransition: {
-            name: "Choose a place ActionNode",
+            name: "Choose a place",
             class: "ActionNode",
             options: [
               {
                 class: "StageNode-StoryItem",
-                name: "city.ogg item",
+                name: "city",
                 audio: "5e6cb53d9acef28a6f57edf317c31bda673d64ed.ogg",
                 image: "5e36825dd492cd428ec8408f55f8dddfe6a1248a.png",
                 audioPath: undefined,
@@ -250,14 +250,14 @@ export const expectedMoyPack: Pack = {
                 imagePath: undefined,
                 imageTimestamp: undefined,
                 okTransition: {
-                  name: "city.ogg ActionNode",
+                  name: "city",
                   class: "ActionNode",
                   options: [
                     {
                       class: "StageNode-Story",
                       audio: "46fe70d98b9763ab70b7e7cea4627e4f8b7c585a.ogg",
                       image: null,
-                      name: "city.ogg Stage node",
+                      name: "city",
                       okTransition: null,
                     },
                   ],
@@ -271,16 +271,16 @@ export const expectedMoyPack: Pack = {
                 audioTimestamp: undefined,
                 imagePath: undefined,
                 imageTimestamp: undefined,
-                name: "alice-jungle.ogg item",
+                name: "alice-jungle",
                 audio: "78adc1006ff121cbf1c7052a02be47c398aecd78.ogg",
                 image: "da5e7052795b59001f09e2caf27412ef8212f23f.png",
                 okTransition: {
-                  name: "alice-jungle.ogg ActionNode",
+                  name: "alice-jungle",
                   class: "ActionNode",
                   options: [
                     {
                       class: "StageNode-Story",
-                      name: "alice-jungle.ogg Stage node",
+                      name: "alice-jungle",
                       audio: "f493d4e986a1278ca7db3f7c65bf8ee32535b2e4.ogg",
                       image: null,
                       okTransition: null,
@@ -319,7 +319,7 @@ export const expectedMoyPackSerialized = {
     },
     {
       id: "23189e83-5f4a-41e5-8f2c-0ce25a1f5b2c",
-      name: "Choose a place ActionNode",
+      name: "Choose a place",
       options: [
         "97bb8423-dfba-4bf3-b7e3-4555c99e6bed",
         "5ab4fedf-081e-43ff-bfe1-b81bb3a893a8",
@@ -331,7 +331,7 @@ export const expectedMoyPackSerialized = {
     },
     {
       id: "5b00e9fc-ae5e-465d-908f-8020aba093b4",
-      name: "alice-jungle.ogg ActionNode",
+      name: "alice-jungle",
       options: ["76e84dff-e0a4-4d8e-94a2-3872c84b9510"],
       position: {
         x: 0,
@@ -340,7 +340,7 @@ export const expectedMoyPackSerialized = {
     },
     {
       id: "5697dee5-155c-485f-9032-fe99eee6c407",
-      name: "city.ogg ActionNode",
+      name: "city",
       options: ["e8b2d354-055a-464a-80d8-9885a06f4ae2"],
       position: {
         x: 0,
@@ -411,7 +411,7 @@ export const expectedMoyPackSerialized = {
         optionIndex: 0,
       },
       image: "da5e7052795b59001f09e2caf27412ef8212f23f.png",
-      name: "alice-jungle.ogg item",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "5b00e9fc-ae5e-465d-908f-8020aba093b4",
         optionIndex: 0,
@@ -438,7 +438,7 @@ export const expectedMoyPackSerialized = {
         optionIndex: 1,
       },
       image: null,
-      name: "alice-jungle.ogg Stage node",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "ID",
         optionIndex: 1,
@@ -465,7 +465,7 @@ export const expectedMoyPackSerialized = {
         optionIndex: 0,
       },
       image: "5e36825dd492cd428ec8408f55f8dddfe6a1248a.png",
-      name: "city.ogg item",
+      name: "city",
       okTransition: {
         actionNode: "5697dee5-155c-485f-9032-fe99eee6c407",
         optionIndex: 0,
@@ -492,7 +492,7 @@ export const expectedMoyPackSerialized = {
         optionIndex: 0,
       },
       image: null,
-      name: "city.ogg Stage node",
+      name: "city",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,
@@ -526,7 +526,7 @@ export const expectedMoyPackNextMenuSerialized = {
     },
     {
       id: "23189e83-5f4a-41e5-8f2c-0ce25a1f5b2c",
-      name: "Choose a place ActionNode",
+      name: "Choose a place",
       options: [
         "97bb8423-dfba-4bf3-b7e3-4555c99e6bed",
         "5ab4fedf-081e-43ff-bfe1-b81bb3a893a8",
@@ -538,7 +538,7 @@ export const expectedMoyPackNextMenuSerialized = {
     },
     {
       id: "5b00e9fc-ae5e-465d-908f-8020aba093b4",
-      name: "alice-jungle.ogg ActionNode",
+      name: "alice-jungle",
       options: ["76e84dff-e0a4-4d8e-94a2-3872c84b9510"],
       position: {
         x: 0,
@@ -547,7 +547,7 @@ export const expectedMoyPackNextMenuSerialized = {
     },
     {
       id: "5697dee5-155c-485f-9032-fe99eee6c407",
-      name: "city.ogg ActionNode",
+      name: "city",
       options: ["e8b2d354-055a-464a-80d8-9885a06f4ae2"],
       position: {
         x: 0,
@@ -618,7 +618,7 @@ export const expectedMoyPackNextMenuSerialized = {
         optionIndex: 0,
       },
       image: "da5e7052795b59001f09e2caf27412ef8212f23f.png",
-      name: "alice-jungle.ogg item",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "5b00e9fc-ae5e-465d-908f-8020aba093b4",
         optionIndex: 0,
@@ -645,7 +645,7 @@ export const expectedMoyPackNextMenuSerialized = {
         optionIndex: 1,
       },
       image: null,
-      name: "alice-jungle.ogg Stage node",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,
@@ -672,7 +672,7 @@ export const expectedMoyPackNextMenuSerialized = {
         optionIndex: 0,
       },
       image: "5e36825dd492cd428ec8408f55f8dddfe6a1248a.png",
-      name: "city.ogg item",
+      name: "city",
       okTransition: {
         actionNode: "5697dee5-155c-485f-9032-fe99eee6c407",
         optionIndex: 0,
@@ -699,7 +699,7 @@ export const expectedMoyPackNextMenuSerialized = {
         optionIndex: 0,
       },
       image: null,
-      name: "city.ogg Stage node",
+      name: "city",
       okTransition: {
         actionNode: "ID",
         optionIndex: 1,
@@ -852,7 +852,7 @@ export const expectedFullPack: Pack = {
           name: "Choose a character",
           okTransition: {
             class: "ActionNode",
-            name: "Choose a character ActionNode",
+            name: "Choose a character",
             options: [
               {
                 class: "StageNode-Menu",
@@ -865,7 +865,7 @@ export const expectedFullPack: Pack = {
                 name: "Alice",
                 okTransition: {
                   class: "ActionNode",
-                  name: "Alice ActionNode",
+                  name: "Alice",
                   options: [
                     {
                       class: "StageNode-Menu",
@@ -878,11 +878,11 @@ export const expectedFullPack: Pack = {
                       name: "Choose a place",
                       okTransition: {
                         class: "ActionNode",
-                        name: "Choose a place ActionNode",
+                        name: "Choose a place",
                         options: [
                           {
                             class: "StageNode-StoryItem",
-                            name: "city.ogg item",
+                            name: "city",
                             audio:
                               "5e6cb53d9acef28a6f57edf317c31bda673d64ed.ogg",
                             audioPath: undefined,
@@ -892,7 +892,7 @@ export const expectedFullPack: Pack = {
                             imagePath: undefined,
                             imageTimestamp: undefined,
                             okTransition: {
-                              name: "city.ogg ActionNode",
+                              name: "city",
                               class: "ActionNode",
                               options: [
                                 {
@@ -900,7 +900,7 @@ export const expectedFullPack: Pack = {
                                   audio:
                                     "46fe70d98b9763ab70b7e7cea4627e4f8b7c585a.ogg",
                                   image: null,
-                                  name: "city.ogg Stage node",
+                                  name: "city",
                                   okTransition: null,
                                 },
                               ],
@@ -910,7 +910,7 @@ export const expectedFullPack: Pack = {
                           },
                           {
                             class: "StageNode-StoryItem",
-                            name: "alice-jungle.ogg item",
+                            name: "alice-jungle",
                             audio:
                               "78adc1006ff121cbf1c7052a02be47c398aecd78.ogg",
                             audioPath: undefined,
@@ -922,7 +922,7 @@ export const expectedFullPack: Pack = {
                             imageTimestamp: undefined,
 
                             okTransition: {
-                              name: "alice-jungle.ogg ActionNode",
+                              name: "alice-jungle",
                               class: "ActionNode",
                               options: [
                                 {
@@ -930,7 +930,7 @@ export const expectedFullPack: Pack = {
                                   audio:
                                     "f493d4e986a1278ca7db3f7c65bf8ee32535b2e4.ogg",
                                   image: null,
-                                  name: "alice-jungle.ogg Stage node",
+                                  name: "alice-jungle",
                                   okTransition: null,
                                 },
                               ],
@@ -959,7 +959,7 @@ export const expectedFullPack: Pack = {
                 name: "Bob",
                 okTransition: {
                   class: "ActionNode",
-                  name: "Bob ActionNode",
+                  name: "Bob",
                   options: [
                     {
                       class: "StageNode-Menu",
@@ -972,11 +972,11 @@ export const expectedFullPack: Pack = {
                       name: "Choose a place",
                       okTransition: {
                         class: "ActionNode",
-                        name: "Choose a place ActionNode",
+                        name: "Choose a place",
                         options: [
                           {
                             class: "StageNode-StoryItem",
-                            name: "bob-city.ogg item",
+                            name: "bob-city",
                             audio:
                               "e260fd4baef8d32d1cb598355ca691e5fe401bc8.ogg",
                             audioPath: undefined,
@@ -988,7 +988,7 @@ export const expectedFullPack: Pack = {
                             imageTimestamp: undefined,
 
                             okTransition: {
-                              name: "bob-city.ogg ActionNode",
+                              name: "bob-city",
                               class: "ActionNode",
                               options: [
                                 {
@@ -996,7 +996,7 @@ export const expectedFullPack: Pack = {
                                   audio:
                                     "90304af3b9d81fb1590ad367df796e791ea78750.ogg",
                                   image: null,
-                                  name: "bob-city.ogg Stage node",
+                                  name: "bob-city",
                                   okTransition: null,
                                 },
                               ],
@@ -1006,7 +1006,7 @@ export const expectedFullPack: Pack = {
                           },
                           {
                             class: "StageNode-StoryItem",
-                            name: "jungle.ogg item",
+                            name: "jungle",
                             audio:
                               "78adc1006ff121cbf1c7052a02be47c398aecd78.ogg",
                             audioPath: undefined,
@@ -1016,7 +1016,7 @@ export const expectedFullPack: Pack = {
                             imagePath: undefined,
                             imageTimestamp: undefined,
                             okTransition: {
-                              name: "jungle.ogg ActionNode",
+                              name: "jungle",
                               class: "ActionNode",
                               options: [
                                 {
@@ -1024,7 +1024,7 @@ export const expectedFullPack: Pack = {
                                   audio:
                                     "084c25abd45f62e50aed566e74ae3317bf844d6c.ogg",
                                   image: null,
-                                  name: "jungle.ogg Stage node",
+                                  name: "jungle",
                                   okTransition: null,
                                 },
                               ],
@@ -1202,7 +1202,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "f13a6d12-f14c-417c-a1b9-0d472a81f71f",
-      name: "Choose a character ActionNode",
+      name: "Choose a character",
       options: [
         "d30ffb97-937c-40e4-87cb-52bb165a6f2c",
         "f45fedec-1bd0-4079-b30f-9ba4dcb53251",
@@ -1214,7 +1214,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "ee081709-807e-411f-b849-3564deeec2c1",
-      name: "Bob ActionNode",
+      name: "Bob",
       options: ["43675af0-4a46-43fd-b499-7b0f50d245ae"],
       position: {
         x: 0,
@@ -1223,7 +1223,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "2ce22c3f-ddc9-4e6f-93d8-09aa7f1c6bda",
-      name: "Choose a place ActionNode",
+      name: "Choose a place",
       options: [
         "f8833e95-2723-472c-ae09-37ae39ddfc40",
         "cfd5a4a4-096c-435b-9d76-f8df327448d2",
@@ -1235,7 +1235,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "ea493af6-e355-4c9b-ad47-f2e905740495",
-      name: "jungle.ogg ActionNode",
+      name: "jungle",
       options: ["66a9993c-5b76-47e6-9145-1b41bf1b3a92"],
       position: {
         x: 0,
@@ -1244,7 +1244,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "7db64f23-1f42-4e03-baa9-9a5beee60c0d",
-      name: "bob-city.ogg ActionNode",
+      name: "bob-city",
       options: ["a16263c0-df61-4dd1-a5a0-af39252ccfa2"],
       position: {
         x: 0,
@@ -1253,7 +1253,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "a9a456f5-3087-4506-adc3-8fdbd1120273",
-      name: "Alice ActionNode",
+      name: "Alice",
       options: ["8c446610-4c51-45da-b58b-7a6d0de7f807"],
       position: {
         x: 0,
@@ -1262,7 +1262,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "ed4794d1-59ea-4024-a25a-c20fae794bb3",
-      name: "Choose a place ActionNode",
+      name: "Choose a place",
       options: [
         "92c20747-0778-4db3-8623-b316c429ce19",
         "3688c876-dd4c-4ec0-bea2-d3dc78149183",
@@ -1274,7 +1274,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "faa61366-848e-4606-ba2a-2b34bb2111ff",
-      name: "alice-jungle.ogg ActionNode",
+      name: "alice-jungle",
       options: ["0d9b1524-1929-4c49-925c-462ed3284e70"],
       position: {
         x: 0,
@@ -1283,7 +1283,7 @@ export const expectedFullPackSerialized: SerializedPack = {
     },
     {
       id: "11e10b4b-c9b7-4f9c-8bc9-bd857466267d",
-      name: "city.ogg ActionNode",
+      name: "city",
       options: ["925387b3-e971-4218-bf4e-cb6dd3c493eb"],
       position: {
         x: 0,
@@ -1405,7 +1405,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 0,
       },
       image: "da5e7052795b59001f09e2caf27412ef8212f23f.png",
-      name: "jungle.ogg item",
+      name: "jungle",
       okTransition: {
         actionNode: "ea493af6-e355-4c9b-ad47-f2e905740495",
         optionIndex: 0,
@@ -1432,7 +1432,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 1,
       },
       image: null,
-      name: "jungle.ogg Stage node",
+      name: "jungle",
       okTransition: {
         actionNode: "ID",
         optionIndex: 1,
@@ -1459,7 +1459,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 0,
       },
       image: "95daa0dc90c3f15300edb39c807e9c9d6acff3f6.png",
-      name: "bob-city.ogg item",
+      name: "bob-city",
       okTransition: {
         actionNode: "7db64f23-1f42-4e03-baa9-9a5beee60c0d",
         optionIndex: 0,
@@ -1486,7 +1486,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 0,
       },
       image: null,
-      name: "bob-city.ogg Stage node",
+      name: "bob-city",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,
@@ -1564,7 +1564,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 0,
       },
       image: "da5e7052795b59001f09e2caf27412ef8212f23f.png",
-      name: "alice-jungle.ogg item",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "faa61366-848e-4606-ba2a-2b34bb2111ff",
         optionIndex: 0,
@@ -1591,7 +1591,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 1,
       },
       image: null,
-      name: "alice-jungle.ogg Stage node",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "ID",
         optionIndex: 1,
@@ -1618,7 +1618,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 0,
       },
       image: "5e36825dd492cd428ec8408f55f8dddfe6a1248a.png",
-      name: "city.ogg item",
+      name: "city",
       okTransition: {
         actionNode: "11e10b4b-c9b7-4f9c-8bc9-bd857466267d",
         optionIndex: 0,
@@ -1645,7 +1645,7 @@ export const expectedFullPackSerialized: SerializedPack = {
         optionIndex: 0,
       },
       image: null,
-      name: "city.ogg Stage node",
+      name: "city",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,
@@ -1679,7 +1679,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "f13a6d12-f14c-417c-a1b9-0d472a81f71f",
-      name: "Choose a character ActionNode",
+      name: "Choose a character",
       options: [
         "d30ffb97-937c-40e4-87cb-52bb165a6f2c",
         "f45fedec-1bd0-4079-b30f-9ba4dcb53251",
@@ -1691,7 +1691,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "ee081709-807e-411f-b849-3564deeec2c1",
-      name: "Bob ActionNode",
+      name: "Bob",
       options: ["43675af0-4a46-43fd-b499-7b0f50d245ae"],
       position: {
         x: 0,
@@ -1700,7 +1700,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "2ce22c3f-ddc9-4e6f-93d8-09aa7f1c6bda",
-      name: "Choose a place ActionNode",
+      name: "Choose a place",
       options: [
         "f8833e95-2723-472c-ae09-37ae39ddfc40",
         "cfd5a4a4-096c-435b-9d76-f8df327448d2",
@@ -1712,7 +1712,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "ea493af6-e355-4c9b-ad47-f2e905740495",
-      name: "jungle.ogg ActionNode",
+      name: "jungle",
       options: ["66a9993c-5b76-47e6-9145-1b41bf1b3a92"],
       position: {
         x: 0,
@@ -1721,7 +1721,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "7db64f23-1f42-4e03-baa9-9a5beee60c0d",
-      name: "bob-city.ogg ActionNode",
+      name: "bob-city",
       options: ["a16263c0-df61-4dd1-a5a0-af39252ccfa2"],
       position: {
         x: 0,
@@ -1730,7 +1730,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "a9a456f5-3087-4506-adc3-8fdbd1120273",
-      name: "Alice ActionNode",
+      name: "Alice",
       options: ["8c446610-4c51-45da-b58b-7a6d0de7f807"],
       position: {
         x: 0,
@@ -1739,7 +1739,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "ed4794d1-59ea-4024-a25a-c20fae794bb3",
-      name: "Choose a place ActionNode",
+      name: "Choose a place",
       options: [
         "92c20747-0778-4db3-8623-b316c429ce19",
         "3688c876-dd4c-4ec0-bea2-d3dc78149183",
@@ -1751,7 +1751,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "faa61366-848e-4606-ba2a-2b34bb2111ff",
-      name: "alice-jungle.ogg ActionNode",
+      name: "alice-jungle",
       options: ["0d9b1524-1929-4c49-925c-462ed3284e70"],
       position: {
         x: 0,
@@ -1760,7 +1760,7 @@ export const expectedFullPackNightSerialized = {
     },
     {
       id: "11e10b4b-c9b7-4f9c-8bc9-bd857466267d",
-      name: "city.ogg ActionNode",
+      name: "city",
       options: ["925387b3-e971-4218-bf4e-cb6dd3c493eb"],
       position: {
         x: 0,
@@ -1891,7 +1891,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 0,
       },
       image: "da5e7052795b59001f09e2caf27412ef8212f23f.png",
-      name: "jungle.ogg item",
+      name: "jungle",
       okTransition: {
         actionNode: "ea493af6-e355-4c9b-ad47-f2e905740495",
         optionIndex: 0,
@@ -1918,7 +1918,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 1,
       },
       image: null,
-      name: "jungle.ogg Stage node",
+      name: "jungle",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,
@@ -1945,7 +1945,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 0,
       },
       image: "95daa0dc90c3f15300edb39c807e9c9d6acff3f6.png",
-      name: "bob-city.ogg item",
+      name: "bob-city",
       okTransition: {
         actionNode: "7db64f23-1f42-4e03-baa9-9a5beee60c0d",
         optionIndex: 0,
@@ -1972,7 +1972,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 0,
       },
       image: null,
-      name: "bob-city.ogg Stage node",
+      name: "bob-city",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,
@@ -2050,7 +2050,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 0,
       },
       image: "da5e7052795b59001f09e2caf27412ef8212f23f.png",
-      name: "alice-jungle.ogg item",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "faa61366-848e-4606-ba2a-2b34bb2111ff",
         optionIndex: 0,
@@ -2077,7 +2077,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 1,
       },
       image: null,
-      name: "alice-jungle.ogg Stage node",
+      name: "alice-jungle",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,
@@ -2104,7 +2104,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 0,
       },
       image: "5e36825dd492cd428ec8408f55f8dddfe6a1248a.png",
-      name: "city.ogg item",
+      name: "city",
       okTransition: {
         actionNode: "11e10b4b-c9b7-4f9c-8bc9-bd857466267d",
         optionIndex: 0,
@@ -2131,7 +2131,7 @@ export const expectedFullPackNightSerialized = {
         optionIndex: 0,
       },
       image: null,
-      name: "city.ogg Stage node",
+      name: "city",
       okTransition: {
         actionNode: "ID",
         optionIndex: 0,