+++ utils/utils_test.ts
@@ -121,6 +121,6 @@ Deno.test("convertToValidFilename", () => {
   assertEquals(convertToValidFilename("aze:rty?uiop"), "aze rty uiop");
   assertEquals(
     convertToValidFilename("aéz&'(rtyèeâî@%:a123"),
-    "aéz  (rtyèeâî   a123",
+    "aéz '(rtyèeâî   a123",
   );
 });